refactor map view
