+++ demo-app/src/pages/platform-devices/MapView/index.js
@@ -1,0 +1,1 @@
+export * from './MapView';
